Remove type_asset erc20 for sdk coins (#326)
this property was added to bridged coins originating from Ethereum or similar chain, but after moving to bridge chains like Axelar, they lose erc20 status. The address property should be required for erc20 tokens, but these are not provided and wouldn't describe the latest stage of the token (only the origin).